Refactor code

+++ service/user/repository.ts
@@ -52,17 +52,22 @@ export class SqlUserRepository extends SearchRepository<User, UserFilter> implem
     const stmts: Statement[] = []
     const stmt = buildToInsert(user, "users", userModel, this.param)
     stmts.push(stmt)
-    insertUserRoles(stmts, user.userId, user.roles, this.param)
+    if (user.roles) {
+      insertUserRoles(stmts, user.userId, user.roles, this.param)
+    }
     return this.db.execBatch(stmts, true)
   }
   update(user: User): Promise<number> {
     const stmts: Statement[] = []
     const stmt = buildToUpdate(user, "users", userModel, this.param)
+    console.log(stmt.query)
     if (!stmt.query) {
       return Promise.resolve(-1)
     }
-    stmts.push({ query: `delete from user_roles where user_id = ${this.param(1)}`, params: [user.userId] })
-    insertUserRoles(stmts, user.userId, user.roles, this.param)
+    if (user.roles) {
+      stmts.push({ query: `delete from user_roles where user_id = ${this.param(1)}`, params: [user.userId] })
+      insertUserRoles(stmts, user.userId, user.roles, this.param)
+    }
     return this.db.execBatch(stmts, true)
   }
   patch(user: User): Promise<number> {
